homerun-direct/backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.5.0
django-extensions==3.2.3
djangorestframework_simplejwt==5.5.0
pytest-django==4.10.0
social-auth-app-django==5.4.3
django-cors-headers==4.7.0
google-auth==2.38.0
pytest-cov==6.0.0
django-jazzmin==3.0.1
django-jazzmin==3.0.1
pytest-mock==3.14.0
python-decouple==3.8
stripe==11.6.0
# pip install stripe -i https://mirrors.aliyun.com/pypi/simple/
psycopg2-binary==2.9.10
django-filter==25.1
redis==5.2.1
django-redis==5.4.0
gunicorn==23.0.0
elasticsearch==8.17.2
elasticsearch-dsl==8.17.1
model_bakery
python-dotenv==1.1.0
factory_boy


